feat: use proper type parsing
On-behalf-of: SAP <matthias.bruns@sap.com>
